Changed fetch and seed
Updated seed file to have all the old stuff it used to have that originally seeded from the csv. Why did it get deleted in the first place? Who knows. I also un-commented the fetch call in QueryPage, changed it from localhost to just /api/v1/tests, and BarGraph now gets sent the result of the fetch in props.